Vectorize logo
The current logo was a clear candidate for vectorizing, being a
collection of geometric shapes. This commit recreates the logo from PSD
to SVG. The SVG uses pre-built Rubik's Cube faces and applies CSS 3D
transformations to make it a cube. Further, the colours are somewhat
changed to be more vibrant, unlike the current blue and yellow.

SVGs (vector) is preferable to raster files for a few reasons, namely
file size and not losing any quality when resized, as well as the
semantic declarative structure of SVG.

Favicons are then generated with Real Favicon Generator.